changed bzero() to memset()
changed bzero() to memset() as bzero is deprecated.

replaced bzero() with memset()

replaced calls to bzero() with memset() as bzero is deprecated.

replaced bcopy() with memmove()

Replaced deprecated bcopy() with memmove(), memmove is guaranteed to
work even if memory areas overlap, where as memcpy is designed so that
the memory areas MUST not overlap.

replaced bzero() with memset()

Replaced bzero() with memset() as bzero is deprecated.

replaced bzero() with memset()

changed bzero() which is deprecated to memset().

replaced deprecated function calls

replaced bzero() with memset() and
bcopy() with memmove().

changed bzero to memset

changed bzero() to memset in #define, bzero deprecated.

replaced bzero() with memset()

replaced bzero() with memset() as bzero is deprecated.

changed function prototype

corrected function prototype for PlatypusAgentReceive to match actual
function later in file.

corrected function prototype and deprecated calls

changed function prototype for SguilRecvAgentMsg() and changed bzero()
to memset() as bzero() is deprecated

replaced bzero() with memset()

replaced bzero() with memset() since bzero() is deprecated.