Make it possible to remove processed files
Up to now it was not possible to simply remove processed files without
moving them elsewhere. Add this possibility by treating the empty string
as a special value to the -a option. (Such value would have little sense
anyway). So using -a "" makes barnyard directly remove the file without
trying to move it to another directory.
While at it, rework the ArchiveFile() function to not return anything,
since its return value it's unused.